+++ node/src/types/kernel/api/block/moveBlock/payload.d.ts
@@ -15,17 +15,20 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-/* directories */
-export * as appendBlock from "./appendBlock";
-export * as deleteBlock from "./deleteBlock";
-export * as getBlockKramdown from "./getBlockKramdown";
-export * as getChildBlocks from "./getChildBlocks";
-export * as insertBlock from "./insertBlock";
-export * as moveBlock from "./moveBlock";
-export * as prependBlock from "./prependBlock";
-export * as updateBlock from "./updateBlock";
-
-/* files */
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getDocInfo from "./getDocInfo";
+/**
+ * Move a block
+ */
+export interface IPayload {
+    /**
+     * Block ID to move
+     */
+    id: string;
+    /**
+     * block ID: insert into this block head
+     */
+    parentID?: string;
+    /**
+     * block ID: insert after this block
+     */
+    previousID?: string;
+}