+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - 添加 API `/api/block/appendBlock` | Add API `/api/block/appendBlock`
 - 添加 API `/api/block/updateBlock` | Add API `/api/block/updateBlock`
 - 添加 API `/api/block/deleteBlock` | Add API `/api/block/deleteBlock`
+- 添加 API `/api/block/moveBlock` | Add API `/api/block/moveBlock`
 
 ## v0.0.5 / 2023-06-30